fw: refactor stream reader to std
